* rename to old_astar - will get removed soon
